requirements.txt
python-telegram-bot==20.0a2
requests==2.28.1